crm/requirements.txt
django==4.2.2
python-dotenv==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.1.0
psycopg2-binary==2.9.6
drf-yasg==1.21.6